requirements.txt
Django>=5.1,<6.0
Pillow>=10.0
pycryptodome>=3.20
cryptography>=42.0
supabase>=2.0
python-dotenv>=1.0
